speechbrain-app/requirements.txt
kafka-python
pyignite==0.6.0
torch==1.10.1
torchaudio==0.10.1
speechbrain
